[core>change] Flip-flop cards and rating view update
The question cards now have two faces, on the front the type of
question is shown along with a brief description and on the back
the list of questions corresponding to that type is displayed.

The action buttons were also updated to start the tour and download
the results, as well as the results download view.